Cleanup for stability

+++ dmtx.h
@@ -304,7 +304,7 @@ extern DmtxVector2 *dmtxVector2Sub(DmtxVector2 *vOut, DmtxVector2 *v1, DmtxVecto
 extern DmtxVector2 *dmtxVector2ScaleBy(DmtxVector2 *v, float s);
 extern DmtxVector2 *dmtxVector2Scale(DmtxVector2 *vOut, DmtxVector2 *v, float s);
 extern float dmtxVector2Cross(DmtxVector2 *v1, DmtxVector2 *v2);
-extern float dmtxVector2Norm(DmtxVector2 *v);
+extern int dmtxVector2Norm(DmtxVector2 *v);
 extern float dmtxVector2Dot(DmtxVector2 *v1, DmtxVector2 *v2);
 extern float dmtxVector2Mag(DmtxVector2 *v);
 extern float dmtxDistanceFromRay2(DmtxRay2 *r, DmtxVector2 *q);
@@ -340,7 +340,7 @@ extern DmtxColor3 *dmtxColor3Sub(DmtxColor3 *vOut, DmtxColor3 *v1, DmtxColor3 *v
 extern DmtxColor3 *dmtxColor3ScaleBy(DmtxColor3 *v, float s);
 extern DmtxColor3 *dmtxColor3Scale(DmtxColor3 *vOut, DmtxColor3 *v, float s);
 extern DmtxColor3 *dmtxColor3Cross(DmtxColor3 *vOut, DmtxColor3 *v1, DmtxColor3 *v2);
-extern float dmtxColor3Norm(DmtxColor3 *v);
+extern int dmtxColor3Norm(DmtxColor3 *v);
 extern float dmtxColor3Dot(DmtxColor3 *v1, DmtxColor3 *v2);
 extern float dmtxColor3Mag(DmtxColor3 *v);
 extern float dmtxDistanceFromRay3(DmtxRay3 *r, DmtxColor3 *q);
